documentation/urls.py
# documentation/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import ArticleViewSet
from . import views

router = DefaultRouter()
router.register(r'articles', ArticleViewSet)

urlpatterns = [
    path('login/', views.login_view, name='login'),
    path('logout/', views.logout_view, name='logout'),
]
