requirements.txt
djangorestframework==<version>